Refactor medicamento endpoints and add user auth dependency
Removed medicamento.py endpoint and merged its functionality into medicamentoxusuario.py. Introduced get_current_user dependency for extracting authenticated user from JWT token. Updated MedicamentoUsuario models and schemas to use float for 'dosis'. Added user extraction from token in user_controller and auth_service. Adjusted main.py router imports and registration accordingly.

+++ app/models/medicamentoUsuario.py
@@ -1,8 +1,7 @@
 """
 Modelo de relación MedicamentoUsuario
 """
-from sqlalchemy import Column, Integer, ForeignKey, DateTime, func
-
+from sqlalchemy import Column, Integer, ForeignKey, DateTime, func, Float
 from app.core.database import Base
 
 
@@ -14,7 +13,7 @@ class MedicamentoUsuario(Base):
     idMedicamento = Column(Integer, ForeignKey("medicamentos.id", ondelete="CASCADE"), nullable=False)
     idUnidad = Column(Integer, ForeignKey("unidades.id", ondelete="CASCADE"), nullable=False)
     
-    dosis = Column(Integer, nullable=False)
+    dosis = Column(Float, nullable=False)
     frecuencia_horas = Column(Integer, nullable=False)
     fecha_inicio = Column(DateTime(timezone=True), nullable=False)
     fecha_fin = Column(DateTime(timezone=True), nullable=False)
